requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pyjwt==2.8.0
cryptography==41.0.3
sqlalchemy==2.0.20
aiohttp==3.8.5
celery==5.3.4
redis==5.0.0
bcrypt==4.0.1
flask-jwt-extended==4.5.2

# Testing dependencies
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-flask==1.2.0

# Security testing specific
python-owasp-zap-v2.4==0.0.20
pyOpenSSL==23.2.0
sslyze==5.1.3
dnspython==2.4.2
pycryptodome==3.19.0
selenium==4.12.0
webdriver-manager==4.0.0

# AI and machine learning
scikit-learn==1.3.0
tensorflow==2.13.0
nltk==3.8.1
transformers==4.32.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
colorama==0.4.6
pyyaml==6.0.1
jsonschema==4.19.0
pytest==7.4.0